add some css

+++ src/components/FlashcardDetail.jsx
@@ -1,17 +1,15 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useMemo, useRef, useState } from "react";
 import styles from "./FlashcardDetail.module.css";
 import { WordItem } from "./WordItem";
 import { getJWTToken } from "../helper_function/authentication";
-import { redirect, useActionData, useLoaderData, useLocation, useNavigate, useSubmit } from "react-router-dom";
+import { useActionData, useLoaderData, useLocation, useNavigate, useSubmit } from "react-router-dom";
 import axios, { all } from "axios";
-import { FlashcardItem } from "./FlashcardItem";
+import { AddWordTemplate } from "./AddWordTemplate";
+import { generateUUIDV4 } from "../helper_function/idManager";
+import { getMaxPage } from "../helper_function/handleInput";
 const dataURL = import.meta.env.VITE_DATA_URL;
 
 const loader = async ({ request, params }) => {
-    // get the collection and all its words
-    // const url = new URL(request.url);
-    // const queryParams = url.searchParams;
-    // console.log(queryParams)
     const token = getJWTToken();
     const response = await axios.get(`${dataURL}/flashcard/${params.id}`, {
         headers: {
@@ -29,8 +27,9 @@ const loader = async ({ request, params }) => {
 const action = async ({ request, params }) => {
     const formData = await request.formData();
     const token = getJWTToken();
-    if (formData.get("_action") === "DELETE") {
-        const response = await axios.delete(`${dataURL}/${params.id}/delete`, {
+    if (formData.get("_action") === "DELETE_COLLECTION") {
+        console.log("ACTION DELETE");
+        const response = await axios.delete(`${dataURL}/flashcard/${params.id}/delete`, {
             headers: {
                 Authorization: `Bearer ${token || "no_token"}`,
             },
@@ -41,29 +40,77 @@ const action = async ({ request, params }) => {
         } else {
             alert("There's something wrong!");
         }
-        return data.result;
+        return { isDeleteSuccessfully: data.result };
+    } else if (formData.get("_action") === "ADD") {
+        console.log("ACTION ADD");
+        const response = await axios.post(`${dataURL}/flashcard/${params.id}/add-word`, formData, {
+            headers: {
+                Authorization: `Bearer ${token || "no_token"}`,
+            },
+        });
+        const data = response.data;
+        if (data.result) {
+            console.log(data.data);
+        } else {
+            alert("There's something wrong!");
+        }
+        return { isAddSuccessfully: data.result };
+    } else if (formData.get("_action") === "DELETE_WORD") {
+        console.log("ACTION DELETE WORD");
+        const wordID = formData.get("wordID");
+        if (wordID) {
+            const response = await axios.delete(`${dataURL}/flashcard/${params.id}/words/${wordID}`, {
+                headers: {
+                    Authorization: `Bearer ${token || "no_token"}`,
+                },
+            });
+            const data = response.data;
+            if (data.result) {
+                console.log(data.data);
+            } else {
+                alert("There's something wrong!");
+            }
+        } else console.log("No wordID");
     }
+    return { isDeleteSuccessfully: false, isAddSuccessfully: false };
 };
 
 const FlashcardDetail = (props) => {
     const { collection } = useLoaderData();
+    const [addWordTemplate, setAddWordTemplate] = useState(false);
+    const firstRender = useRef(true);
     const [allWords, setAllWords] = useState([]);
+    const requesting = useRef(false);
     const [currentWords, setCurrentWords] = useState([]);
     const [currentPage, setCurrentPage] = useState(1);
     const [limit, setLimit] = useState(10);
+    const [noOfPracticeWords, setNoOfPracticeWords] = useState(10);
     const submit = useSubmit();
     const currentURL = useLocation();
-    const deleteResult = useActionData();
+    const actionData = useActionData();
     const navigate = useNavigate();
     const [maxPage, setMaxPage] = useState(1);
 
-    const requestWords = async () => {
+    const requestWords = async (isReset) => {
+        console.log("requestWords");
         const ranges = []; // always has even number of elements
-        let start = (currentPage - 1) * limit;
+        let curPage = currentPage;
+        if (currentPage > getMaxPage(collection.noOfWords, limit)) {
+            // if currentPage is greater than the maxPage
+            curPage--;
+            setCurrentPage((state) => state - 1);
+        }
+        let start = (curPage - 1) * limit;
         let end = Math.min(start + limit, collection.noOfWords);
         let isStart = false;
+        const wordArr = firstRender.current || isReset ? new Array(collection.noOfWords).fill(null) : allWords;
+        console.log("wordArr");
+        console.log(wordArr);
+        firstRender.current = false; // unmark the first render
+
+        // find missing words to request
         for (let i = start; i < end; i++) {
-            if (!allWords[i]) {
+            if (!wordArr[i]) {
                 if (!isStart) {
                     ranges.push(i);
                     isStart = true;
@@ -83,6 +130,7 @@ const FlashcardDetail = (props) => {
         for (let i = 1; i <= ranges.length / 2; i++) {
             queryString += `start${i}=${ranges[(i - 1) * 2]}&limit${i}=${ranges[(i - 1) * 2 + 1]}`;
         }
+        console.log("ranges");
         if (ranges.length != 0) {
             const result = await axios.get(`${dataURL}/flashcard/${collection.id}/words/${queryString}`, {
                 headers: {
@@ -92,7 +140,7 @@ const FlashcardDetail = (props) => {
             const data = result.data;
             if (data.result) {
                 const words = data.data;
-                const newAllWords = [...allWords];
+                const newAllWords = [...wordArr];
                 for (const entry of Object.entries(words)) {
                     console.log(entry);
                     if (entry[1].result) {
@@ -107,47 +155,66 @@ const FlashcardDetail = (props) => {
             setDisplayWords();
             console.log("Data's available already");
         }
+        requesting.current = false;
     };
 
     useEffect(() => {
-        if (deleteResult) {
+        if (actionData?.isDeleteSuccessfully) {
             navigate("/flashcard", { replace: true });
         }
-    }, [deleteResult]);
+    }, [actionData]);
 
     useEffect(() => {
         if (collection) {
             if (currentPage === 1) {
                 // if currentPage is already 1, try request here
-                requestWords();
+                if (!requesting.current) {
+                    console.log("request in limit");
+                    requesting.current = true;
+                    requestWords(false);
+                }
             } else {
                 // if currentPage is not 1
                 setCurrentPage(1); // return to the first page
             }
-            setMaxPage(Math.floor(collection.noOfWords / limit) + 1); // re-evaluate the max page based on current limit
+            setMaxPage(getMaxPage(collection.noOfWords, limit)); // re-evaluate the max page based on current limit
         }
     }, [limit]);
 
     useEffect(() => {
         if (collection) {
-            requestWords();
+            setAddWordTemplate(false); // close the add word template when turn page
+            if (!requesting.current) {
+                console.log("request in currentPage");
+                requesting.current = true;
+                requestWords(false);
+            }
         }
     }, [currentPage]);
 
     useEffect(() => {
         if (collection) {
-            setMaxPage(Math.floor(collection.noOfWords / limit) + 1);
-            setAllWords(Array(collection.noOfWords).fill(null));
+            console.log("collection change");
+            setMaxPage(getMaxPage(collection.noOfWords, limit));
+            if (!requesting.current) {
+                console.log("request in collection");
+                requesting.current = true;
+                requestWords(true);
+            }
         }
     }, [collection]);
 
     useEffect(() => {
+        console.log("allWords change");
         setDisplayWords();
     }, [allWords]);
 
     const setDisplayWords = () => {
         if (allWords.length > 0) {
+            console.log("allWords");
+            console.log(allWords);
             const displayWords = allWords.slice((currentPage - 1) * limit, currentPage * limit);
+            console.log("displayWords");
             console.log(displayWords);
             setCurrentWords(displayWords);
         }
@@ -167,13 +234,47 @@ const FlashcardDetail = (props) => {
     };
 
     const deleteThisCollection = () => {
-        submit(
-            { _action: "DELETE" },
-            {
-                method: "DELETE",
-                action: currentURL.pathname,
-            }
-        );
+        const result = window.confirm("Do you actually want to delete this collection?");
+        if (result) {
+            submit(
+                { _action: "DELETE_COLLECTION" },
+                {
+                    method: "DELETE",
+                    action: currentURL.pathname,
+                }
+            );
+        }
+    };
+
+    const deleteWord = (wordID) => {
+        const result = window.confirm("Do you actually want to delete this word?");
+        if (result && wordID) {
+            submit(
+                { _action: "DELETE_WORD", wordID },
+                {
+                    method: "DELETE",
+                    action: currentURL.pathname,
+                }
+            );
+        }
+    };
+
+    const addWordToDatabase = (word) => {
+        const formData = new FormData();
+        formData.append("_action", "ADD");
+        formData.append("id", generateUUIDV4());
+        formData.append("collectionID", collection.id);
+        formData.append("word", word.word);
+        formData.append("definition", word.definition);
+        formData.append("example", word.example);
+        submit(formData, {
+            method: "POST",
+        });
+    };
+
+    const goToLearnCollection = () => {
+        if (collection) navigate(`/flashcard/${collection.id}/practice/?limit=${noOfPracticeWords}`);
+        else alert("Please try again");
     };
 
     const generateControlButton = () => {
@@ -230,9 +331,52 @@ const FlashcardDetail = (props) => {
             <div className={styles["container"]}>
                 <div>{collection?.title || "NULL TITLE"}</div>
                 <div>{collection?.description || ""}</div>
+                <button
+                    className={styles["learn-button"]}
+                    onClick={() => {
+                        goToLearnCollection();
+                    }}
+                >
+                    Learn
+                </button>
+                <label htmlFor="practice-words-select">
+                    Choose the number of words you want to practice
+                    <select
+                        id="practice-words-select"
+                        value={noOfPracticeWords}
+                        onChange={(e) => setNoOfPracticeWords(Number(e.target.value))}
+                    >
+                        <option value="10">10 words</option>
+                        <option value="20">20 words</option>
+                        <option value="30">30 words</option>
+                    </select>
+                </label>
+
+                <div className={styles["option-container"]}>
+                    <button
+                        className={`${styles["add-button"]} ${addWordTemplate && styles["invisible"]}`}
+                        onClick={() => {
+                            setAddWordTemplate(true);
+                        }}
+                    >
+                        Add word
+                    </button>
+                    <button
+                        className={`${styles["add-button"]} ${addWordTemplate || styles["invisible"]}`}
+                        onClick={() => {
+                            setAddWordTemplate(false);
+                        }}
+                    >
+                        Done
+                    </button>
+                    <button className={styles["delete-button"]} onClick={() => deleteThisCollection()}>
+                        Delete
+                    </button>
+                </div>
+
                 <select
+                    className={styles["limit-select"]}
                     value={limit}
-                    name="noOfWords"
                     onChange={(e) => {
                         console.log(e.target.value);
                         setLimit(Number(e.target.value));
@@ -242,10 +386,10 @@ const FlashcardDetail = (props) => {
                     <option value={20}>20 words/page</option>
                     <option value={50}>50 words/page</option>
                 </select>
-                <button className={styles["delete-button"]} onClick={() => deleteThisCollection()}>
-                    Delete
-                </button>
-                {currentWords && currentWords.length > 0 && currentWords.map((item) => <WordItem word={item} />)}
+                <AddWordTemplate active={addWordTemplate} addWord={addWordToDatabase} />
+                {currentWords &&
+                    currentWords.length > 0 &&
+                    currentWords.map((item) => <WordItem word={item} deleteWord={deleteWord} />)}
                 <div className={styles["page-controller"]}>{generateControlButton()}</div>
             </div>
         </>
